Made a few `Enums` methods obsolete
The methods were refactored into new properties in the new `Enums<TEnum>` class, which should make calls to them cleaner (as there are no parens, and it is impossible to infer the generic type of the methods anyways).

In addition, the new properties return immutable arrays that can be shared safely by all callers to the value or atomic value getters the entire lifetime of the program, removing the need to duplicate the array multiple times.